Setting up main page layout
The tabs will be generated via js, but the
baseline is the same so that'll be done using
regular html, which is almost done.

One bug left is that the border on the buttons is
not uniform. Must also set a class to change
the bottom of the selected button (so it seems
like you're on that tab) that can then be set via
the main js.